Pull reference handling to separate file
