+++ archive.py
@@ -11,8 +11,7 @@ import re
 import dateutil.parser as dp
 import shutil
 import warnings
-from types import SimpleNamespace
-from collections import namedtuple
+import archive_reference
 
 parser = argparse.ArgumentParser(description="Archive repo issues and PRs.")
 parser.add_argument("repo", help="GitHub repo to archive (e.g. quicwg/base-drafts)")
@@ -700,114 +699,19 @@ def getPRs(refFile, fields=gql_PullRequest_Fields, updateOld=False):
                 get_more_issues = False
 
 
-def newReferenceFile():
-    return SimpleNamespace(
-        magic=current_magic,
-        issues=dict(),
-        prs=dict(),
-        issues_only=True,
-        lastSuccess=None,
-        canCopy=False,
-        filename=None,
-    )
-
-
-def loadReference(filename):
-    fileIsValid = False
-    reference = newReferenceFile()
-    reference.filename = filename
-
-    if filename:
-        try:
-            with open(filename, "r") as ref_file:
-                raw_reference = json.load(ref_file)
-
-            fileIsValid = True
-            for element in ("magic", "timestamp", "issues", "repo"):
-                if element not in raw_reference:
-                    fileIsValid = False
-                    break
-
-            if fileIsValid and (
-                raw_reference["magic"] == current_magic
-                or raw_reference["magic"] in upgrades
-            ):
-                reference.magic = raw_reference["magic"]
-            else:
-                warnings.warn("Input file does not appear to be generated by this tool")
-                fileIsValid = False
-
-            if fileIsValid and raw_reference["repo"] != args.repo:
-                warnings.warn("Input file was generated from a different repo")
-                fileIsValid = False
-
-            if fileIsValid:
-                reference.lastSuccess = dp.parse(raw_reference["timestamp"])
-                reference.issues = dict(
-                    [(issue["number"], issue) for issue in raw_reference["issues"]]
-                )
-                if "pulls" in raw_reference:
-                    reference.issues_only = False
-                    reference.prs = dict(
-                        [(pr["number"], pr) for pr in raw_reference["pulls"]]
-                    )
-                else:
-                    reference.issues_only = True
-                    reference.prs = dict()
-
-        except:
-            warnings.warn("Unable to read input file; proceeding without it")
-            pass
-
-    if not fileIsValid:
-        return newReferenceFile()
-
-    reference.canCopy = bool(reference.issues) and bool(
-        reference.prs or args.issuesOnly
-    )
-
-    return upgradeReference(reference)
-
-
-#########################
-## Upgrade definitions ##
-#########################
-
-current_magic = "E!vIA5L86J2I"
-
-UpgradeInstruction = namedtuple("UpgradeInstruction", ["result", "issues", "prs"])
-
-upgrades = {
-    "B8n2c@e8kvfx": UpgradeInstruction(
-        result="E!vIA5L86J2I",
-        issues=None,
-        prs="""
-            fragment prFields on PullRequest {
-                number
-                baseRepository { nameWithOwner }
-                baseRefName
-                baseRefOid
-                headRepository { nameWithOwner }
-                headRefName
-                headRefOid
-                mergeCommit { oid }
-            }""",
-    )
-}
-
-
 def upgradeReference(reference):
     try:
-        while reference.magic in upgrades:
-            if upgrades[reference.magic].issues:
-                getIssues(reference, upgrades[reference.magic].issues, True)
-            if upgrades[reference.magic].prs:
-                getPRs(reference, upgrades[reference.magic].prs, True)
-            reference.magic = upgrades[reference.magic].result
+        while archive_reference.is_legacy_magic(reference.magic):
+            instructions = archive_reference.get_upgrade_instructions(reference.magic)
+            if instructions.issues:
+                getIssues(reference, instructions.issues, True)
+            if instructions.prs:
+                getPRs(reference, instructions.prs, True)
+            reference.magic = instructions.result
     except:
         pass
 
-    if reference.magic != current_magic:
+    if not archive_reference.is_current_magic(reference.magic):
         warnings.warn(
             "Unable to upgrade input file to current version; proceeding without it"
         )
@@ -823,7 +727,10 @@ def upgradeReference(reference):
 (owner, repo) = args.repo.split("/", 1)
 
 ## Read in the reference files, if any
-reference = loadReference(args.refFile)
+reference = upgradeReference(archive_reference.loadReference(args.refFile, args.repo))
+reference.canCopy = bool(reference.issues) and bool(
+    reference.prs or args.issuesOnly
+)
 
 ## Download from GitHub the full issues list (if no reference) or the updated issues list (if reference)
 getIssues(reference)
@@ -857,7 +764,7 @@ if reference.canCopy and args.outFile:
     shutil.copyfile(args.refFile, args.outFile)
 else:
     output = {
-        "magic": current_magic,
+        "magic": archive_reference.get_current_magic(),
         "timestamp": now.isoformat(),
         "repo": args.repo,
         "labels": labels_ref,
